acmicpc
acmicpc

+++ 5347.cpp
@@ -1,26 +1,26 @@
 #include "stdio.h"
-long a,b;
-
-long gcd(long x, long y){
-	long t = x;
+long long a,b;
+ 
+long long gcd(long long x, long long y){
+	long long t = x;
 	x = y;
 	y = t % y;
-
+ 
 	if(y == 0){
 		return x;
 	}
 	return gcd(x,y);
 }
-
+ 
 int main(){
-	int num;
-	scanf("%d",&num);
-
+	long long num;
+	scanf("%lld",&num);
+ 
 	for(int i = 0; i < num; i++){
-		scanf("%ld %ld", &a, &b);
-		printf("%ld\n", (a*b) /gcd(a,b));
-
+		scanf("%lld %lld", &a, &b);
+		printf("%lld\n", (a*b) /gcd(a,b));
+ 
 	}
-
+ 
 	return 0;
 }
